validate quiz route

+++ server/src/routes/quiz.js
@@ -2,6 +2,7 @@ import { Router } from "express";
 import { ensureLoggedIn } from "connect-ensure-login";
 import Quiz from "../models/quiz.js";
 import User from "../models/user.js";
+import { CATEGORIES, QUIZ_STATUSES } from "../utils/constants.js";
 
 const quizRouter = Router();
 
@@ -16,10 +17,16 @@ quizRouter.get("/", async (req, res) => {
 });
 
 quizRouter.get("/user", ensureLoggedIn(), async (req, res) => {
-  const quizzes = await Quiz.find({ createdBy: req.user.id }, null, {
-    sort: { createdAt: "desc" },
-  });
-  res.json({ data: quizzes });
+  const user = await User.findById(req.user.id);
+
+  if (user) {
+    const quizzes = await Quiz.find({ createdBy: user.id }, null, {
+      sort: { createdAt: "desc" },
+    });
+    return res.json({ data: quizzes });
+  }
+
+  res.status(404).send("No user found with that ID");
 });
 
 quizRouter.get("/:id", async (req, res) => {
@@ -32,13 +39,15 @@ quizRouter.get("/:id", async (req, res) => {
 });
 
 quizRouter.delete("/:id", ensureLoggedIn(), async (req, res) => {
+  const user = await User.findById(req.user.id);
+  if (!user) return res.status(404).send("No user found with that ID");
+
   const quiz = await Quiz.findOneAndDelete({
     _id: req.params.id,
-    createdBy: req.user.id,
+    createdBy: user.id,
   }).exec();
 
   if (quiz) {
-    const user = await User.findById(req.user.id);
     user.quizzes = user.quizzes.filter((_id) => quiz.id != _id);
     await user.save();
   }
@@ -49,12 +58,22 @@ quizRouter.patch("/:id", ensureLoggedIn(), async (req, res) => {
   const { title, description, surveySchema, category, image } = req.body;
   const props = { title, description, surveySchema, category, image };
 
+  const user = await User.findById(req.user.id);
+  if (!user) return res.status(404).send("No user found with that ID");
+
   const quiz = await Quiz.findOne({
     _id: req.params.id,
-    createdBy: req.user.id,
+    createdBy: user.id,
   });
   Object.keys(props).forEach(
-    (prop) => props[prop] && (quiz[prop] = props[prop])
+    (prop) => {
+      if (props[prop]) {
+        if (prop == "category" && !CATEGORIES.includes(props[prop])) return;
+        if (prop == "status" && !QUIZ_STATUSES.includes(props[prop])) return;
+
+        quiz[prop] = props[prop];
+      }
+    }
   );
   await quiz.save();
 
@@ -63,6 +82,8 @@ quizRouter.patch("/:id", ensureLoggedIn(), async (req, res) => {
 
 quizRouter.post("/:id/like", ensureLoggedIn(), async (req, res) => {
   const user = await User.findById(req.user.id);
+  if (!user) return res.status(404).send("No user found with that ID");
+
   const quiz = await Quiz.findOne({
     _id: req.params.id,
   });
@@ -82,20 +103,23 @@ quizRouter.post("/:id/like", ensureLoggedIn(), async (req, res) => {
 });
 
 quizRouter.post("/", ensureLoggedIn(), async (req, res) => {
-  const { title, description, surveySchema, category, image, status } =
+  let { title, description, surveySchema, category, image, status } =
     req.body;
+
+  if (!CATEGORIES.includes(category)) category = "misc";
+  if (!QUIZ_STATUSES.includes(status)) status = "drafted";
 
   const newQuiz = new Quiz({
     title,
     description,
     surveySchema,
-    category,
     image,
     status,
+    category,
     createdBy: req.user.id,
   });
-  const quiz = await newQuiz.save();
   const user = await User.findById(req.user.id);
+  const quiz = await newQuiz.save();
 
   if (user) {
     user.quizzes.push(quiz.id);
@@ -103,7 +127,7 @@ quizRouter.post("/", ensureLoggedIn(), async (req, res) => {
     return res.status(201).json({ data: quiz });
   }
 
-  res.status(404).send("User Not Found");
+  res.status(404).send("No user found with that ID");
 });
 
 export default quizRouter;
